move TestConfig_TrustCACerts test
