+++ test/e2e/kappcontroller/http_test.go
@@ -123,121 +123,3 @@ spec:
 		}
 	})
 }
-
-func TestHTTPSSelfSignedCerts(t *testing.T) {
-	env := e2e.BuildEnv(t)
-	logger := e2e.Logger{}
-	kapp := e2e.Kapp{t, env.Namespace, logger}
-	sas := e2e.ServiceAccounts{env.Namespace}
-
-	// When updating, certs and keys must be regenerated for server and added to server.go and config-test/config-map.yml
-	serverNamespace := "https-server"
-
-	yaml1 := fmt.Sprintf(`---
-apiVersion: kappctrl.k14s.io/v1alpha1
-kind: App
-metadata:
-  name: test-https
-  annotations:
-    kapp.k14s.io/change-group: kappctrl-e2e.k14s.io/apps
-spec:
-  serviceAccountName: kappctrl-e2e-ns-sa
-  fetch:
-  - http:
-      url: https://https-svc.%s.svc.cluster.local:443/deployment.yml
-  template:
-  - ytt: {}
-  deploy:
-  - kapp:
-      inspect: {}
-      intoNs: %s
-`, serverNamespace, env.Namespace) + sas.ForNamespaceYAML()
-
-	name := "test-https"
-	httpsServerName := "test-https-server"
-
-	cleanUp := func() {
-		kapp.Run([]string{"delete", "-a", name})
-		kapp.Run([]string{"delete", "-a", httpsServerName, "-n", serverNamespace})
-	}
-
-	cleanUp()
-	defer cleanUp()
-
-	logger.Section("deploy https server with self signed certs", func() {
-		kapp.Run([]string{"deploy", "-f", "../assets/https-server/server.yml", "-f", "../assets/https-server/certs-for-custom-ca.yml", "-a", httpsServerName})
-	})
-
-	logger.Section("deploy", func() {
-		kapp.RunWithOpts([]string{"deploy", "-f", "-", "-a", name},
-			e2e.RunOpts{IntoNs: true, StdinReader: strings.NewReader(yaml1)})
-
-		out := kapp.Run([]string{"inspect", "-a", name, "--raw", "--tty=false", "--filter-kind=App"})
-
-		var cr v1alpha1.App
-
-		err := yaml.Unmarshal([]byte(out), &cr)
-		if err != nil {
-			t.Fatalf("Failed to unmarshal: %s", err)
-		}
-
-		expectedStatus := v1alpha1.AppStatus{
-			GenericStatus: v1alpha1.GenericStatus{
-				Conditions: []v1alpha1.Condition{{
-					Type:   v1alpha1.ReconcileSucceeded,
-					Status: corev1.ConditionTrue,
-				}},
-				ObservedGeneration:  1,
-				FriendlyDescription: "Reconcile succeeded",
-			},
-			Deploy: &v1alpha1.AppStatusDeploy{
-				ExitCode: 0,
-				Finished: true,
-			},
-			Fetch: &v1alpha1.AppStatusFetch{
-				ExitCode: 0,
-			},
-			Inspect: &v1alpha1.AppStatusInspect{
-				ExitCode: 0,
-			},
-			Template: &v1alpha1.AppStatusTemplate{
-				ExitCode: 0,
-			},
-			ConsecutiveReconcileSuccesses: 1,
-		}
-
-		{
-			// deploy
-			if !strings.Contains(cr.Status.Deploy.Stdout, "Wait to:") {
-				t.Fatalf("Expected non-empty deploy output: '%s'", cr.Status.Deploy.Stdout)
-			}
-			cr.Status.Deploy.StartedAt = metav1.Time{}
-			cr.Status.Deploy.UpdatedAt = metav1.Time{}
-			cr.Status.Deploy.Stdout = ""
-
-			// fetch
-			if !strings.Contains(cr.Status.Fetch.Stdout, "kind: LockConfig") {
-				t.Fatalf("Expected non-empty fetch output: '%s'", cr.Status.Fetch.Stdout)
-			}
-			cr.Status.Fetch.StartedAt = metav1.Time{}
-			cr.Status.Fetch.UpdatedAt = metav1.Time{}
-			cr.Status.Fetch.Stdout = ""
-
-			// inspect
-			if !strings.Contains(cr.Status.Inspect.Stdout, "Resources in app 'test-https-ctrl'") {
-				t.Fatalf("Expected non-empty inspect output: '%s'", cr.Status.Inspect.Stdout)
-			}
-			cr.Status.Inspect.UpdatedAt = metav1.Time{}
-			cr.Status.Inspect.Stdout = ""
-
-			// template
-			cr.Status.Template.UpdatedAt = metav1.Time{}
-			cr.Status.Template.Stderr = ""
-		}
-
-		if !reflect.DeepEqual(expectedStatus, cr.Status) {
-			t.Fatalf("Status is not same: %#v vs %#v", expectedStatus, cr.Status)
-		}
-	})
-
-}
